fix: forceSync in case of auto save
