+++ web/app/[workspaceSlug]/(projects)/projects/(detail)/[projectId]/pages/(detail)/[pageId]/page.tsx
@@ -1,45 +1,29 @@
 "use client";
 
-import { useRef, useState } from "react";
 import { observer } from "mobx-react";
 import Link from "next/link";
 import { useParams } from "next/navigation";
 import useSWR from "swr";
-// document-editor
-import { EditorRefApi, useEditorMarkings } from "@plane/editor";
-// types
-import { TPage } from "@plane/types";
 // ui
-import { TOAST_TYPE, getButtonStyling, setToast } from "@plane/ui";
+import { getButtonStyling } from "@plane/ui";
 // components
 import { LogoSpinner } from "@/components/common";
 import { PageHead } from "@/components/core";
 import { IssuePeekOverview } from "@/components/issues";
-import { PageEditorBody, PageEditorHeaderRoot } from "@/components/pages";
 // helpers
 import { cn } from "@/helpers/common.helper";
 // hooks
 import { usePage, useProjectPages } from "@/hooks/store";
-import { useAppRouter } from "@/hooks/use-app-router";
-import { usePageDescription } from "@/hooks/use-page-description";
+import { PageRoot } from "./page-root";
 
 const PageDetailsPage = observer(() => {
-  // states
-  const [sidePeekVisible, setSidePeekVisible] = useState(window.innerWidth >= 768 ? true : false);
-  const [editorReady, setEditorReady] = useState(false);
-  const [readOnlyEditorReady, setReadOnlyEditorReady] = useState(false);
-  // refs
-  const editorRef = useRef<EditorRefApi>(null);
-  const readOnlyEditorRef = useRef<EditorRefApi>(null);
-  // router
-  const router = useAppRouter();
   const { workspaceSlug, projectId, pageId } = useParams();
+
   // store hooks
-  const { createPage, getPageById } = useProjectPages();
+  const { getPageById } = useProjectPages();
   const page = usePage(pageId?.toString() ?? "");
-  const { access, description_html, id, name } = page;
-  // editor markings hook
-  const { markings, updateMarkings } = useEditorMarkings();
+  const { id, name } = page;
+
   // fetch page details
   const { error: pageDetailsError } = useSWR(
     pageId ? `PAGE_DETAILS_${pageId}` : null,
@@ -48,16 +32,6 @@ const PageDetailsPage = observer(() => {
       revalidateIfStale: false,
       revalidateOnFocus: false,
       revalidateOnReconnect: false,
-    }
-  );
-
-  // project-description
-  const { handleDescriptionChange, isDescriptionReady, pageDescriptionYJS, handleSaveDescription } = usePageDescription(
-    {
-      editorRef,
-      page,
-      projectId,
-      workspaceSlug,
     }
   );
 
@@ -84,56 +58,12 @@ const PageDetailsPage = observer(() => {
       </div>
     );
 
-  const handleCreatePage = async (payload: Partial<TPage>) => await createPage(payload);
-
-  const handleDuplicatePage = async () => {
-    const formData: Partial<TPage> = {
-      name: "Copy of " + name,
-      description_html: editorRef.current?.getHTML() ?? description_html ?? "<p></p>",
-      access,
-    };
-
-    await handleCreatePage(formData)
-      .then((res) => router.push(`/${workspaceSlug}/projects/${projectId}/pages/${res?.id}`))
-      .catch(() =>
-        setToast({
-          type: TOAST_TYPE.ERROR,
-          title: "Error!",
-          message: "Page could not be duplicated. Please try again later.",
-        })
-      );
-  };
-
   return (
     <>
       <PageHead title={name} />
       <div className="flex h-full flex-col justify-between">
         <div className="h-full w-full flex-shrink-0 flex flex-col overflow-hidden">
-          <PageEditorHeaderRoot
-            editorRef={editorRef}
-            readOnlyEditorRef={readOnlyEditorRef}
-            editorReady={editorReady}
-            readOnlyEditorReady={readOnlyEditorReady}
-            handleDuplicatePage={handleDuplicatePage}
-            handleSaveDescription={handleSaveDescription}
-            markings={markings}
-            page={page}
-            sidePeekVisible={sidePeekVisible}
-            setSidePeekVisible={(state) => setSidePeekVisible(state)}
-          />
-          <PageEditorBody
-            editorRef={editorRef}
-            handleEditorReady={(val) => setEditorReady(val)}
-            readOnlyEditorRef={readOnlyEditorRef}
-            handleReadOnlyEditorReady={() => setReadOnlyEditorReady(true)}
-            markings={markings}
-            page={page}
-            sidePeekVisible={sidePeekVisible}
-            updateMarkings={updateMarkings}
-            handleDescriptionChange={handleDescriptionChange}
-            isDescriptionReady={isDescriptionReady}
-            pageDescriptionYJS={pageDescriptionYJS}
-          />
+          <PageRoot page={page} projectId={projectId.toString()} workspaceSlug={workspaceSlug.toString()} />
           <IssuePeekOverview />
         </div>
       </div>
